refactor(organization-detail-page): modularize organization details into Overview, Settings, and Hierarchy tabs for improved structure and maintainability
